Rambley's eye markings are not dependent on eyebrows
The createRambleyEyeMarkings method no longer takes in the ellipse for Rambley's eyebrows, instead taking in the bounds of the head. Additionally, both the eyebrows and the ellipse for the eye markings are based off the bottom of the head instead of the top, since the bottom y-coordinate doesn't change when the ears are added to the head.