Undid "More accurate Rambley snout"
I just realized that the reference I was using for Rambley "without eyes" probably still had eyes, and what I was interpreting as his mask-like markings was probably the outline of his closed eye. As such, I'm undoing most of the work done for the "More accurate Rambley snout" commit, as that work was done in error.