Fang side flag is now flipped flag
I've changed the flag used to control whether Rambley's fang is on the left or right side of his face to be a more applicable flag, now called the flipped flag. This still is used to control which side of Rambley's face his fang will appear on, but it will also be used to control the side of other elements too, such as the scarf.