Top head ellipse is now reused
The snout Ellipse2D was found to only be used within the paintRambleySnout method, and I cannot foresee it being used for much else. As such, the ellipse2 scratch Ellipse2D object is used instead. There is now a headEllipse Ellipse2D object used to create the top part of Rambley's head. While this is only used once, it is expected to be used in future components.