Border and Shadow are now separate
I've changed it so that there are two separate flags for the border around Rambley (now referred to as the outline) and Rambley's shadow, allowing them to be independently controlled. I've also renamed the color constants with the word OUTLINE in their name to instead just have LINE in their name, and the RAMBLEY_BORDER_COLOR constant is now RAMBLEY_OUTLINE_COLOR. Additionally the RAMBLEY_BORDER_THICKNESS constant is now RAMBLEY_OUTLINE_THICKNESS, getRambleyOutlineStroke is now getRambleyLineStroke, getRambleyBorderStroke is now getRambleyOutlineStroke and paintRambleyBorderAndShadow is now paintRambleyOutlineAndShadow. Also, the paintRambleyOutlineAndShadow is now called all the time, but it only engages if either the outline or the drop shadow are to be painted, and will only paint them if they are to be painted. I've also temporarily removed the scarf from being painted.