Update RambleyPainter.java
I've renamed the RAMBLEY_EAR_MULTIPLIER constant to RAMBLEY_ANIMALWAVE_MULTIPLIER, since it may be used for more than just the ear equations in the future, as other equations provided by AnimalWave may be used in the future. The INTERNAL_RENDER_WIDTH and INTERNAL_RENDER_HEIGHT constants have been made private, at least temporarily, as they may be removed and replaced with a more dynamic system later on. A constant has been added that is used to offset the upper part of the neck portion of Rambley's scarf in order to get the lower part. A first of possibly multiple dedicated CubicCurve2D objects for the scarf, named scarfCurve1, has been added.