Neck part of Scarf Finished
I've moved the code for generating the neck portion of Rambley's scarf into a method called createRambleyNeckScarf, and the code for generating the detail for the neck part of the scarf to a method called createRambleyNeckScarfDetail. I've also added a method called paintRambleyNeckScarf for painting the neck part of the scarf.