Print code is in GeometryMath + debug method docs
I've moved the print test code to the GeometryMath library, as it may be used for future debugging and may be worth keeping around. I've also added documentation to the remaining debug methods that state that these methods will be removed at some point. There are also three new debug methods, which are used to calculate control points.