paintRambley clips graphics context
The paintRambley method now clips the graphics context to prevent it from painting outside the area it is given. As a result, the paintRambleyOutlineAndShadow method now no longer takes in the bounds of the area to paint in and no longer clips the graphics context it renders to, as that was meant to prevent it from painting outside the bounds.