Head shape is now truly centered
The shape produced by the getRambleyEarlessHead method is now more centered in the area for it. Previously, due to how the earless head shape was produced, it would be about 8.913x10^-9 off center, which while imperceptible to the human eye, would often result in rounding issues that were perceptible later down in the code when not accounted for. While the code still accounts for the head shape not being centered, it now is centered.