Internal Size and Position change if scarf is shown
I've made it so that, when there is no scarf, Rambley is rendered at 256x256 with a y-coordinate of 70, and when Rambley is wearing his scarf, Rambley is rendered at 320x320 with a y-coordinate of 80. This is a temporary solution to altering the size and position of Rambley until I think of a better solution to controlling the size and position of Rambley.