Reworked the lower part of the scarf
I've reworked the code that creates the lower part of the scarf to use two fitting quadratic bezier curves instead of interpolating the curves. I've also made it so that it is using the scarf6 object as the path, at least for now. Additionally, I've added a note to the ear generating method that it uses all 8 scratch Point2D objects.