Glitchy Rambley Head
I've made it so that the earless portion of Rambley's head now becomes more angular when Rambley is glitching out. I've also added a reference to the methods for getting and setting whether Rambley is glitchy to the documentation of some of the methods affected when Rambley is glitching out.