RambleyPainter uses Rhombus2D for Background
I've made it so that instead of creating the background polka dot rhombuses using a Path2D, it instead uses a Rhombus2D so that the code for moving it around can be shared with the circle background polka dots. I've also removed the getBackgroundDotRhombus method and removed the Path2D and Ellipse2D parameters from getBackgroundDot. The background now uses dedicated Rhombus2D and Ellipse2D objects. There is also a generic scratch Rhombus2D object, but it is commented out as nothing uses it yet.