Update RambleyTestViewer.java
I've added a check for existing files when saving a file using the test program. This will cause a prompt to pop up asking the user if they want to replace the file or not, and depending on the user's choice, either the file will be replaced, the user will be prompted to enter a different file, or the file will not be saved.