Create TwoAxisSlider.java
I've added a component that shows two sliders for the x and y axis, along with a chart displaying where the x and y are. It's still in development, and the sliders may be removed at a later time, but currently it works. There is no way currently to set or retrieve values at the moment, though that will be added later.