Fixed getting the intersection of two lines
I've fixed the getLineIntersection method (the one that takes in two lines). Previously, it used the Line2D.linesIntersect method to try to determine if the two lines intersect, however it turns out that that method checks if two line SEGMENTS intersect, instead of whether two infinitely long lines intersect at some point. I also realized that the only way the two lines would not intersect is if they were either parallel or coincident lines, and in those cases the denominator would be zero.