Re-added the background and pixel grid painters
I've moved the RambleyBackgroundPainter and RambleyPixelGridPainter painters into the test folder since I have plans to use them, however I need to think about how they are going to be used in the final program. I've done some minor updates to them to reflect some of the updates made to the code in RambleyPainter (that code will likely be removed from RambleyPainter and delegated to those two painters in the future), but not everything is finished for those two painters.