Starting adding pref node to Rambley4J
I've added some keys for the preference node in Rambley4J, along with adding the preference node name. I've also added the node and a debug mode value, however the preference node is not initialized yet. The debug mode is set at the construction of the JFrame (it will be set by the main method when given a debug argument.