Updated saving
I've added a png file extension filter to the save file chooser. I've reworked the saving code to check if the png file filter is active and the file entered does not have the file extension, and if so, adds the png file extension to the file. I've also made it so that the current directory for the save file chooser is saved and loaded from the preference node.