0 Line spacing results in a rectangle filled
If the line spacing of the pixel grid is set to zero, then a rectangle is filled in the area instead of the pixel grid, since we don't need to calculate a line for every pixel in the image. When the line spacing is 0, there is no space between the lines, so the calculations for the pixel grid's path break anyway.