Background dots are now scaled
The background dots are now scaled. I took the lazy way to do it for the background polka-dots (technically it's not lazy, as Rambley is scaled in the same way, but in this case, since there's only 2 fields like with the pixel grid as opposed to a ton of fixed values, it is lazy).